% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcb.R
\docType{data}
\name{dcb}
\alias{dcb}
\title{Percentage of molting feathers in a 9 primarie bird}
\format{
A data frame with 47 rows and 2 variables:
\describe{
  \item{P}{name of the feather}
  \item{molt}{percentage of molting (between 0 and 1)}
  ...
}
}
\usage{
dcb
}
\description{
A dataset containing the percentage of molting feathers in a 9 primarie bird
}
\keyword{datasets}
