% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Percentage of molting feathers in an only 10 primaries bird}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{P}{name of the feather}
  \item{molt}{percentage of molting (between 0 and 1)}
  ...
}
}
\usage{
df
}
\description{
A dataset containing the percentage of molting feathers in an only 10 primaries bird
}
\keyword{datasets}
