% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcf.R
\docType{data}
\name{fcf}
\alias{fcf}
\title{Percentage of molting feathers in a 10 primarie bird}
\format{
A data frame with 48 rows and 2 variables:
\describe{
  \item{P}{name of the feather}
  \item{molt}{percentage of molting (between 0 and 1)}
  ...
}
}
\usage{
fcf
}
\description{
A dataset containing the percentage of molting feathers in a 10 primarie bird
}
\keyword{datasets}
