% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moult_color.R
\name{moult_color}
\alias{moult_color}
\title{{moult_color}}
\usage{
moult_color(i, data)
}
\arguments{
\item{i}{a value to run the loop}

\item{data}{a data table to choose the color}
}
\value{
color of the feather
}
\description{
moult_color
}
