% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarie_10.R
\name{primarie_10}
\alias{primarie_10}
\title{{primarie_10}}
\usage{
primarie_10(i, data)
}
\arguments{
\item{i}{a value to run the loop}

\item{data}{a data table to choose the color}
}
\value{
a graph
}
\description{
primarie_10
}
