% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarie_9.R
\name{primarie_9}
\alias{primarie_9}
\title{{primarie_9}}
\usage{
primarie_9(i, data)
}
\arguments{
\item{i}{a value to run the loop}

\item{data}{a data table to choose the color}
}
\value{
a graph
}
\description{
primarie_9
}
