% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moult_color.R
\name{moult_color}
\alias{moult_color}
\title{moult_color}
\arguments{
\item{i}{a value to run the loop}

\item{data}{a data table to choose the color}
}
\value{
color of the feather depending of the percentage, called for side effect
}
\description{
moult_color
}
