% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primarie_9.R
\name{primarie_9}
\alias{primarie_9}
\title{{primarie_9}}
\arguments{
\item{i}{a value to run the loop}

\item{data}{a data table to choose the color}
}
\value{
No return value but etablish a molt graph for a 9 primaries' bird
}
\description{
primarie_9
}
\examples{
primarie_9(2,dcb)



}
